1) Fixed price type(Double -> BigDecimal) 2) Fixed OverflowStrategy 3) Created exception classes to handle exceptions during requesting and parsing response

+++ src/main/scala/stockmonitoringbot/stockpriceservices/CurrencyExchangeRateInfo.scala
@@ -9,5 +9,5 @@ case class CurrencyExchangeRateInfo(from: String,
                                     descriptionFrom: String,
                                     to: String,
                                     descriptionTo: String,
-                                    rate: Double,
+                                    rate: BigDecimal,
                                     lastRefreshed: ZonedDateTime)
